[master] refactor on repository message consumer
