+++ tests/unit/master/test_repository.py
@@ -859,381 +859,3 @@ class RepositoryRevisionTest(TestCase):
     def test_check_skip_dont_skip(self):
         self.rev.body = 'some body\n'
         self.assertTrue(self.rev.create_builds())
-
-
-class RepositoryMessageConsumerTest(TestCase):
-
-    def tearDown(self):
-        repository.RepositoryMessageConsumer._stop_consuming_messages = False
-
-    @patch.object(repository.Repository, 'get', AsyncMagicMock())
-    @patch.object(repository.RepositoryRevision, 'objects', Mock())
-    @async_test
-    async def test_add_builds(self):
-        repo = create_autospec(spec=repository.Repository,
-                               mock_cls=AsyncMagicMock)
-        repo.build_manager = AsyncMagicMock()
-        repository.Repository.get.return_value = repo
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'asdf',
-                    'revisions_ids': []}
-        to_list = AsyncMagicMock()
-        repository.RepositoryRevision\
-                  .objects.filter.return_value.to_list = to_list
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer._add_builds(msg)
-        self.assertTrue(repository.Repository.get.called)
-        self.assertTrue(to_list.called)
-        self.assertTrue(msg.acknowledge.called)
-
-    @patch.object(repository.Repository, 'get', AsyncMagicMock())
-    @patch.object(repository.RepositoryRevision, 'objects', Mock())
-    @patch.object(repository.utils, 'log', Mock())
-    @async_test
-    async def test_add_builds_exception(self):
-        repo = create_autospec(spec=repository.Repository,
-                               mock_cls=AsyncMagicMock)
-        repository.Repository.get.return_value = repo
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'asdf', 'revisions_ids': []}
-        to_list = AsyncMagicMock(side_effect=Exception)
-        repository.RepositoryRevision\
-                  .objects.filter.return_value.to_list = to_list
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer._add_builds(msg)
-        self.assertTrue(repository.Repository.get.called)
-        self.assertTrue(to_list.called)
-        self.assertTrue(msg.acknowledge.called)
-
-    @patch.object(repository.Repository, 'get', AsyncMagicMock(
-        side_effect=repository.Repository.DoesNotExist))
-    @patch.object(repository.RepositoryRevision, 'objects', Mock())
-    @patch.object(repository.utils, 'log', Mock())
-    @async_test
-    async def test_add_builds_repo_dont_exist(self):
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'asdf'}
-        to_list = AsyncMagicMock()
-        repository.RepositoryRevision\
-                  .objects.filter.return_value.to_list = to_list
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer._add_builds(msg)
-        self.assertTrue(repository.Repository.get.called)
-        self.assertFalse(to_list.called)
-        self.assertTrue(msg.acknowledge.called)
-
-    @patch.object(repository.Repository, 'get', AsyncMagicMock())
-    @patch.object(repository.Slave, 'objects', Mock())
-    @async_test
-    async def test_add_requested_build(self):
-        repo = create_autospec(spec=repository.Repository,
-                               mock_cls=AsyncMagicMock)
-        repository.Repository.get.return_value = repo
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'asdf',
-                    'branch': 'master',
-                    'slaves_ids': ['some-slave-id']}
-        to_list = AsyncMagicMock()
-        repository.Slave.objects.filter.return_value.to_list = to_list
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer._add_requested_build(msg)
-        self.assertTrue(repo.start_build.called)
-
-    @patch.object(repository.Repository, 'get', AsyncMagicMock(
-        return_value=None))
-    @patch.object(repository.Slave, 'objects', Mock())
-    @async_test
-    async def test_add_requested_build_no_repo(self):
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'asdf',
-                    'branch': 'master',
-                    'slave_ids': ['some-slave-id']}
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer._add_requested_build(msg)
-        self.assertFalse(repository.Slave.objects.filter.called)
-
-    @patch.object(repository.Repository, 'get', AsyncMagicMock())
-    @patch.object(repository.Slave, 'objects', Mock())
-    @async_test
-    async def test_add_requested_build_no_slaves(self):
-        repo = create_autospec(spec=repository.Repository,
-                               mock_cls=AsyncMagicMock)
-        repository.Repository.get.return_value = repo
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'asdf',
-                    'branch': 'master'}
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer._add_requested_build(msg)
-        self.assertTrue(repo.start_build.called)
-
-    @patch.object(repository.Repository, 'get', AsyncMagicMock())
-    @patch.object(repository.utils, 'log', Mock())
-    @async_test
-    async def test_add_requested_build_exception(self):
-        repo = create_autospec(spec=repository.Repository,
-                               mock_cls=AsyncMagicMock)
-        repository.Repository.get.return_value = repo
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'asdf'}
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer._add_requested_build(msg)
-        self.assertFalse(repo.start_build.called)
-        self.assertTrue(repository.utils.log.called)
-
-    @patch.object(repository.RepositoryMessageConsumer, '_add_builds',
-                  AsyncMagicMock())
-    @patch.object(repository.revisions_added, 'consume', AsyncMagicMock())
-    @async_test
-    async def test_wait_revisions(self):
-
-        consumer = repository.revisions_added.consume.return_value
-
-        async def fm(cancel_on_timeout=False):
-            repository.RepositoryMessageConsumer.\
-                _stop_consuming_messages = True
-
-        consumer.fetch_message = fm
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer.wait_revisions()
-        self.assertTrue(message_consumer._add_builds.called)
-
-    @patch.object(repository.RepositoryMessageConsumer, '_add_builds',
-                  AsyncMagicMock())
-    @patch.object(repository.revisions_added, 'consume', AsyncMagicMock())
-    @async_test
-    async def test_wait_revisions_timeout(self):
-
-        consumer = repository.revisions_added.consume.return_value
-
-        async def fm(cancel_on_timeout=False):
-            repository.RepositoryMessageConsumer.\
-                _stop_consuming_messages = True
-            raise repository.ConsumerTimeout
-
-        consumer.fetch_message = fm
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer.wait_revisions()
-        self.assertFalse(message_consumer._add_builds.called)
-
-    @patch.object(repository.repo_notifications, 'consume',
-                  AsyncMagicMock(spec=repository.repo_notifications.consume))
-    @patch.object(repository.RepositoryMessageConsumer, '_add_requested_build',
-                  AsyncMagicMock(
-                      spec=repository.RepositoryMessageConsumer.
-                      _add_requested_build))
-    @async_test
-    async def test_wait_build_requests(self):
-        consumer = repository.repo_notifications.consume.return_value
-
-        async def fm(cancel_on_timeout=False):
-            repository.RepositoryMessageConsumer.\
-                _stop_consuming_messages = True
-
-        consumer.fetch_message = fm
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer.wait_build_requests()
-        self.assertTrue(
-            repository.RepositoryMessageConsumer._add_requested_build.called)
-
-    @patch.object(repository.repo_notifications, 'consume',
-                  AsyncMagicMock(spec=repository.repo_notifications.consume))
-    @patch.object(repository.RepositoryMessageConsumer, '_add_requested_build',
-                  AsyncMagicMock(
-                      spec=repository.RepositoryMessageConsumer.
-                      _add_requested_build))
-    @async_test
-    async def test_wait_build_requests_timeout(self):
-        consumer = repository.repo_notifications.consume.return_value
-
-        async def fm(cancel_on_timeout=False):
-            repository.RepositoryMessageConsumer.\
-                _stop_consuming_messages = True
-            raise repository.ConsumerTimeout
-
-        consumer.fetch_message = fm
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer.wait_build_requests()
-        self.assertFalse(
-            repository.RepositoryMessageConsumer._add_requested_build.called)
-
-    @patch.object(repository.RepositoryMessageConsumer, '_get_repo_from_msg',
-                  AsyncMagicMock(
-                      spec=repository.RepositoryMessageConsumer.
-                      _get_repo_from_msg,
-                      return_value=None))
-    @async_test
-    async def test_remove_repo_no_repo(self):
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'some-id'}
-        message_consumer = repository.RepositoryMessageConsumer()
-        r = await message_consumer._remove_repo(msg)
-        self.assertFalse(r)
-
-    @patch.object(repository.RepositoryMessageConsumer, '_get_repo_from_msg',
-                  AsyncMagicMock(spec=repository.RepositoryMessageConsumer.
-                                 _get_repo_from_msg,
-                                 return_value=create_autospec(
-                                     spec=repository.Repository,
-                                     mock_cls=AsyncMagicMock)))
-    @async_test
-    async def test_remove_repo(self):
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'some-id'}
-        message_consumer = repository.RepositoryMessageConsumer()
-        r = await message_consumer._remove_repo(msg)
-        repo = message_consumer._get_repo_from_msg.\
-            return_value
-        self.assertTrue(repo.remove.called)
-        self.assertTrue(r)
-
-    @patch.object(repository.RepositoryMessageConsumer, '_get_repo_from_msg',
-                  AsyncMagicMock(spec=repository.RepositoryMessageConsumer.
-                                 _get_repo_from_msg,
-                                 return_value=create_autospec(
-                                     spec=repository.Repository,
-                                     mock_cls=AsyncMagicMock)))
-    @patch.object(repository.utils, 'log', Mock())
-    @async_test
-    async def test_remove_repo_exception(self):
-        repo = repository.RepositoryMessageConsumer._get_repo_from_msg.\
-            return_value
-        repo.remove.side_effect = Exception
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'some-id'}
-        message_consumer = repository.RepositoryMessageConsumer()
-        r = await message_consumer._remove_repo(msg)
-        self.assertTrue(repo.remove.called)
-        self.assertTrue(repository.utils.log.called)
-        self.assertTrue(r)
-
-    @patch.object(repository.RepositoryMessageConsumer, '_get_repo_from_msg',
-                  AsyncMagicMock(spec=repository.RepositoryMessageConsumer.
-                                 _get_repo_from_msg,
-                                 return_value=create_autospec(
-                                     spec=repository.Repository,
-                                     mock_cls=AsyncMagicMock)))
-    @patch.object(repository.Repository, 'update_code', AsyncMagicMock(
-        spec=repository.Repository.update_code))
-    @async_test
-    async def test_update_repo(self):
-        repo = repository.RepositoryMessageConsumer.\
-            _get_repo_from_msg.return_value
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'some-id'}
-        message_consumer = repository.RepositoryMessageConsumer()
-        r = await message_consumer._update_repo(msg)
-        self.assertTrue(r)
-        self.assertTrue(repo.update_code.called)
-
-    @patch.object(repository.RepositoryMessageConsumer, '_get_repo_from_msg',
-                  AsyncMagicMock(spec=repository.RepositoryMessageConsumer.
-                                 _get_repo_from_msg,
-                                 return_value=create_autospec(
-                                     spec=repository.Repository,
-                                     mock_cls=AsyncMagicMock)))
-    @patch.object(repository.Repository, 'update_code', AsyncMagicMock(
-        spec=repository.Repository.update_code))
-    @patch.object(repository.utils, 'log', Mock())
-    @async_test
-    async def test_update_repo_exception(self):
-        repo = repository.RepositoryMessageConsumer._get_repo_from_msg.\
-            return_value
-        repo.update_code.side_effect = Exception
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'some-id'}
-        message_consumer = repository.RepositoryMessageConsumer()
-        r = await message_consumer._update_repo(msg)
-        self.assertTrue(r)
-        self.assertTrue(repo.update_code.called)
-        self.assertTrue(repository.utils.log.called)
-
-    @patch.object(repository.RepositoryMessageConsumer, '_get_repo_from_msg',
-                  AsyncMagicMock(spec=repository.RepositoryMessageConsumer.
-                                 _get_repo_from_msg,
-                                 return_value=None))
-    @patch.object(repository.Repository, 'update_code', AsyncMagicMock(
-        spec=repository.Repository.update_code))
-    @async_test
-    async def test_update_repo_no_repo(self):
-        msg = AsyncMagicMock()
-        msg.body = {'repository_id': 'some-id'}
-        message_consumer = repository.RepositoryMessageConsumer()
-        r = await message_consumer._update_repo(msg)
-        self.assertFalse(r)
-
-    @patch.object(repository.repo_notifications, 'consume',
-                  AsyncMagicMock(spec=repository.repo_notifications.consume))
-    @patch.object(repository.RepositoryMessageConsumer, '_remove_repo',
-                  AsyncMagicMock(
-                      spec=repository.RepositoryMessageConsumer._remove_repo))
-    @async_test
-    async def test_wait_removal_request(self):
-        consumer = repository.repo_notifications.consume.return_value
-
-        async def fm(cancel_on_timeout=False):
-            repository.RepositoryMessageConsumer.\
-                _stop_consuming_messages = True
-            return AsyncMagicMock()
-
-        consumer.fetch_message = fm
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer.wait_removal_request()
-        kw = repository.repo_notifications.consume.call_args[1]
-        self.assertTrue(message_consumer._remove_repo.called)
-        self.assertEqual(kw['routing_key'], 'repo-removal-requested')
-
-    @patch.object(repository.repo_notifications, 'consume',
-                  AsyncMagicMock(spec=repository.repo_notifications.consume))
-    @patch.object(repository.RepositoryMessageConsumer, '_remove_repo',
-                  AsyncMagicMock(spec=repository.RepositoryMessageConsumer.
-                                 _remove_repo))
-    @async_test
-    async def test_wait_removal_request_timeout(self):
-        consumer = repository.repo_notifications.consume.return_value
-
-        async def fm(cancel_on_timeout=False):
-            repository.RepositoryMessageConsumer.\
-                _stop_consuming_messages = True
-            raise repository.ConsumerTimeout
-
-        consumer.fetch_message = fm
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer.wait_removal_request()
-        self.assertFalse(message_consumer._remove_repo.called)
-
-    @patch.object(repository.repo_notifications, 'consume',
-                  AsyncMagicMock(spec=repository.repo_notifications.consume))
-    @patch.object(repository.RepositoryMessageConsumer, '_update_repo',
-                  AsyncMagicMock(
-                      spec=repository.RepositoryMessageConsumer._update_repo))
-    @async_test
-    async def test_wait_upate_request(self):
-
-        async def fm(cancel_on_timeout=False):
-            repository.RepositoryMessageConsumer.\
-                _stop_consuming_messages = True
-            msg = AsyncMagicMock()
-            msg.body = {}
-            return msg
-
-        consumer = repository.repo_notifications.consume.return_value
-        consumer.fetch_message = fm
-        message_consumer = repository.RepositoryMessageConsumer()
-        await message_consumer.wait_update_request()
-
-    def test_run(self):
-        message_consumer = repository.RepositoryMessageConsumer()
-        message_consumer.wait_revisions = AsyncMagicMock(
-            spec=message_consumer.wait_revisions)
-        message_consumer.wait_build_requests = AsyncMagicMock(
-            spec=message_consumer.wait_build_requests)
-        message_consumer.wait_removal_request = AsyncMagicMock(
-            spec=message_consumer.wait_removal_request)
-        message_consumer.wait_update_request = AsyncMagicMock(
-            spec=message_consumer.wait_update_request)
-
-        message_consumer.run()
-        self.assertTrue(all([message_consumer.wait_revisions.called,
-                             message_consumer.wait_build_requests.called,
-                             message_consumer.wait_revisions.called,
-                             message_consumer.wait_update_request.called]))
