+++ toxicbuild/master/consumers.py
@@ -20,6 +20,9 @@ import asyncio
 import signal
 from asyncamqp.exceptions import ConsumerTimeout
 from toxicbuild.core.utils import LoggerMixin
+from toxicbuild.master.exchanges import repo_notifications, revisions_added
+from toxicbuild.master.repository import Repository, RepositoryRevision
+from toxicbuild.master.slave import Slave
 
 
 __doc__ = """This module implements consumers for messages published
@@ -31,26 +34,27 @@ class BaseConsumer(LoggerMixin):
     """A base class for the consumer that react to incomming messages
     from exchanges"""
 
-    def __init__(self, exchange, msg_callback, loop=None):
+    def __init__(self, exchange, msg_callback, routing_key=None, loop=None):
         """:param exchange: The exchange in which messages are published
           for the consumer to fetch.
-        :param msg_callback: A callable that receives the message"""
+        :param msg_callback: A callable that receives the message.
+        :param routing_key: The routing key to consume messages.
+        """
 
         self.exchange = exchange
         self.msg_callback = msg_callback
+        self.routing_key = routing_key
         self.loop = loop or asyncio.get_event_loop()
         self._stop = False
         self._running_tasks = 0
         signal.signal(signal.SIGTERM, self.sync_shutdown)
 
-    async def run(self, routing_key=None):
-        """Starts the consumer.
-
-        :param routing_key: The routing key to consume messages"""
+    async def run(self):
+        """Starts the consumer."""
 
         kw = {'timeout': 1000}
-        if routing_key:
-            kw['routing_key'] = routing_key
+        if self.routing_key:
+            kw['routing_key'] = self.routing_key
 
         async with await self.exchange.consume(**kw) as consumer:
             while not self._stop:
@@ -78,3 +82,141 @@ class BaseConsumer(LoggerMixin):
 
     def sync_shutdown(self, signum=None, frame=None):
         self.loop.run_until_complete(self.shutdown())
+
+
+class RepositoryMessageConsumer(LoggerMixin):
+    """Waits for messages published in the ``repo_notifications``
+    exchange and reacts to them."""
+
+    REPOSITORY_CLASS = Repository
+    REPOSITORY_REVISION_CLASS = RepositoryRevision
+    SLAVE_CLASS = Slave
+
+    revision_consumer = None
+    build_consumer = None
+    removal_consumer = None
+    update_consumer = None
+
+    def __init__(self):
+        type(self).revision_consumer = BaseConsumer(revisions_added,
+                                                    self._add_builds)
+        type(self).build_consumer = BaseConsumer(repo_notifications,
+                                                 self._add_requested_build,
+                                                 routing_key='build-requested')
+        type(self).removal_consumer = BaseConsumer(
+            repo_notifications, self._remove_repo,
+            routing_key='repo-removal-requested')
+        type(self).update_consumer = BaseConsumer(
+            repo_notifications, self._update_repo,
+            routing_key='update-code-requested')
+
+    def run(self):
+        """Starts the repository message consumer.
+        """
+        self.log('[init] Waiting revisions')
+        asyncio.ensure_future(type(self).revision_consumer.run())
+        self.log('[init] Waiting build requests')
+        asyncio.ensure_future(type(self).build_consumer.run())
+        self.log('[init] Waiting removal requests')
+        asyncio.ensure_future(type(self).removal_consumer.run())
+        self.log('[init] Waiting code update requests')
+        asyncio.ensure_future(type(self).update_consumer.run())
+
+    async def _get_repo_from_msg(self, msg):
+        try:
+            repo = await self.REPOSITORY_CLASS.get(
+                id=msg.body['repository_id'])
+        except Repository.DoesNotExist:
+            log_msg = '[_get_repo_from_msg] repo {} does not exist'.format(
+                msg.body['repository_id'])
+            self.log(log_msg, level='warning')
+
+            await msg.acknowledge()
+            return
+
+        return repo
+
+    async def _add_builds(self, msg):
+        repo = await self._get_repo_from_msg(msg)
+        if not repo:
+            return
+
+        body = msg.body
+        try:
+            revisions = await self.REPOSITORY_REVISION_CLASS.objects.filter(
+                id__in=body['revisions_ids']).to_list()
+
+            await repo.build_manager.add_builds(revisions)
+        except Exception as e:
+            log_msg = 'Error adding builds for repo {}. '.format(repo.id)
+            log_msg += 'Exception was {}'.format(str(e))
+            self.log(log_msg, level='error')
+        finally:
+            await msg.acknowledge()
+
+    async def _add_requested_build(self, msg):
+        repo = await self._get_repo_from_msg(msg)
+        if not repo:
+            return
+
+        body = msg.body
+        try:
+            branch = body['branch']
+            builder_name = body.get('builder_name')
+            named_tree = body.get('named_tree')
+            slaves_ids = body.get('slaves_ids')
+            if slaves_ids:
+                slaves = await self.SLAVE_CLASS.objects.filter(
+                    id__in=body['slaves_ids']).to_list()
+            else:
+                slaves = []
+
+            await repo.start_build(branch, builder_name=builder_name,
+                                   named_tree=named_tree, slaves=slaves)
+        except Exception as e:
+            log_msg = 'Error starting builds for {}. '.format(repo.id)
+            log_msg += 'Exception was {}'.format(str(e))
+            self.log(log_msg, level='error')
+        finally:
+            await msg.acknowledge()
+
+    async def _remove_repo(self, msg):
+        await msg.acknowledge()
+        repo = await self._get_repo_from_msg(msg)
+        if not repo:
+            return False
+        try:
+            await repo.remove()
+        except Exception as e:
+            log_msg = '[_remove_repo] Error removing repo {}'.format(repo.id)
+            log_msg += '\nOriginal exception was {}'.format(str(e))
+            self.log(log_msg, level='error')
+
+        return True
+
+    async def _update_repo(self, msg):
+        await msg.acknowledge()
+        repo = await self._get_repo_from_msg(msg)
+        repo_branches = msg.body.get('repo_branches')
+        external = msg.body.get('external')
+        wait_for_lock = msg.body.get('wait_for_lock', False)
+        kw = {'repo_branches': repo_branches,
+              'external': external,
+              'wait_for_lock': wait_for_lock}
+        if not repo:
+            return False
+        try:
+            await repo.update_code(**kw)
+        except Exception as e:
+            log_msg = '[_update_repo] Error update repo {}'.format(repo.id)
+            log_msg += '\nOriginal exception was {}'.format(str(e))
+            self.log(log_msg, level='error')
+
+        return True
+
+    @classmethod
+    def stop(cls):
+        cls.revision_consumer.stop()
+        cls.build_consumer.stop()
+        cls.removal_consumer.stop()
+        cls.update_consumer.stop()
